+++ csv.hpp
@@ -14,5 +14,5 @@ using namespace std;
 vector<vector<double>> csv_reader(string fname);
 
 //fonction pour récupérer les données et avoir un nouveau fichier 
-void csv_write(vector<int> parcours);
+void csv_write(vector<double> parcours);
 #endif
